requirements.txt
tornado==6.0.3
timeago==1.0.8
motor==2.1.0
apscheduler==3.5.3
aioredis==1.1.0
lxml
aiocache==0.10.1
msgpack==0.5.6
opencc-python-reimplemented==0.1.5
w3lib==1.21.0
pyquery==1.4.1
elasticsearch==7.6.0
aioelasticsearch==0.6.0
elasticsearch-dsl==7.1.0
aiomysql==0.0.20
ujson==2.0.3
requests==2.23.0
pymongo==3.10.1
pymysql==0.9.2